+++ manim_editor/editor/__init__.py
@@ -5,4 +5,5 @@ from .load_project import *
 from .commands import *
 from .config import *
 from .edit_project import *
-from .manim_assets import IconNormal, IconSkip, IconLoop,IconCompleteLoop
+from .assets import IconNormal, IconSkip, IconLoop, IconCompleteLoop
+
